wip: minor refactoring
