+++ features/id-check-wrapper/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/idcheckwrapper/internal/biometrictoken/BiometricTokenResult.kt
@@ -1,0 +1,19 @@
+package uk.gov.onelogin.criorchestrator.features.idcheckwrapper.internal.biometrictoken
+
+import uk.gov.idcheck.repositories.api.vendor.BiometricToken
+
+sealed interface BiometricTokenResult {
+    object Loading : BiometricTokenResult
+
+    data class Success(
+        val token: BiometricToken,
+    ) : BiometricTokenResult
+
+    object Offline : BiometricTokenResult
+
+    data class Error(
+        val message: String,
+        val statusCode: Int? = null,
+        val error: Exception? = null,
+    ) : BiometricTokenResult
+}
