+++ features/id-check-wrapper/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/idcheckwrapper/internal/biometrictoken/network/BiometricTokenProviderImplTest.kt
@@ -95,8 +95,6 @@ class BiometricTokenProviderImplTest {
 
             val encoded = json.encodeToString(successData)
 
-            println(encoded)
-
             whenever(
                 biometricApi.getBiometricToken(
                     SESSION_ID,
